feat(utils/php): optimize namespace retrieval and template building
This commit optimizes the namespace retrieval in the `get_namespace`
function by reading and parsing the composer.json file once, instead of
reading it line by line. It also improves the template building functions
by adding a new `build_empty` function and refactoring the existing
functions to use a helper function for getting the template file. The
`build_by_content` function is also updated to use the new `build_empty`
function when necessary. A new empty PHP template is added as well.